+++ examples/buttons.rs
@@ -1,4 +1,3 @@
-use textual::demo_snapshot::{SnapshotArgs, snapshot_widget};
 use textual::prelude::*;
 
 struct ButtonsApp;
@@ -60,15 +59,5 @@ async fn main() -> Result<()> {
     if cfg!(test) {
         return Ok(());
     }
-
-    if let Some(args) = SnapshotArgs::parse() {
-        let widget = build_buttons_widget();
-        return snapshot_widget(
-            &widget,
-            &args,
-            Some(std::path::Path::new("examples/button.tcss")),
-        );
-    }
-
-    run_textual_app(ButtonsApp).await
+    run_textual_app_or_snapshot(ButtonsApp).await
 }